# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/teachers/{id}/acesso": {
      "post": {
        "operationId": "TeachersController_gerarAcesso",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfessorComSenhaTemporariaResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "teachers"
        ]
      }
    },
...
    "/api/v1/companies": {
      "get": {
        "operationId": "CompaniesController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmpresaPaginadaResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "companies"
        ]
      },
      "post": {
        "operationId": "CompaniesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmpresaCriadaResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "companies"
        ]
      }
    },
    "/api/v1/companies/{id}": {
      "get": {
        "operationId": "CompaniesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmpresaResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "companies"
        ]
      },
      "patch": {
        "operationId": "CompaniesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmpresaResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "companies"
        ]
      }
    },
    "/api/v1/companies/{id}/admins": {
      "get": {
        "operationId": "CompaniesController_listAdmins",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminDaEmpresaResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "companies"
        ]
      }
    },
    "/api/v1/companies/{id}/admins/{usuarioId}/senha-temporaria": {
      "post": {
        "operationId": "CompaniesController_gerarSenhaDeAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "usuarioId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SenhaDeAdminResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "companies"
        ]
      }
    },
    "/api/v1/companies/{id}/status": {
      "patch": {
        "operationId": "CompaniesController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmpresaResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "companies"
        ]
      }
    },
    "/api/v1/public/companies/{slug}": {
      "get": {
        "operationId": "PublicCompaniesController_porSlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmpresaPublicaResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "public"
        ]
      }
    },
...
    "/api/v1/me/teacher/classes/{id}/ocorrencias": {
      "get": {
        "operationId": "MeTeacherAttendanceController_ocorrencias",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dias",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcorrenciasDaTurmaPaginadasResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/teacher/attendance/{ocupacaoId}": {
      "get": {
        "operationId": "MeTeacherAttendanceController_chamada",
        "parameters": [
          {
            "name": "ocupacaoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChamadaResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      },
      "put": {
        "operationId": "MeTeacherAttendanceController_salvar",
        "parameters": [
          {
            "name": "ocupacaoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalvarChamadaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChamadaSalvaResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/teacher/attendance/{ocupacaoId}/nao-houve": {
      "put": {
        "operationId": "MeTeacherAttendanceController_naoHouve",
        "parameters": [
          {
            "name": "ocupacaoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChamadaNaoHouveResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
...
  },
  "info": {
    "title": "PlayCK API",
    "description": "Contrato da API do back — fonte para geração de tipos nos 3 frontends (ADR-001)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProfessorComSenhaTemporariaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "companyId": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string",
            "example": "Carlos Lima"
          },
          "telefone": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ativo",
              "inativo"
            ]
          },
          "usuarioId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "fotoUrl": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "senhaTemporaria": {
            "type": "string",
            "example": "Xk4p-9Qm2"
          }
        },
        "required": [
          "id",
          "companyId",
          "nome",
          "telefone",
          "email",
          "status",
          "usuarioId",
          "fotoUrl",
          "createdAt",
          "senhaTemporaria"
        ]
      },
...
      "EmpresaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string",
            "example": "Smart Tennis"
          },
          "slug": {
            "type": "string",
            "example": "smart-tennis"
          },
          "esportes": {
            "example": [
              "Tênis",
              "Padel"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ativa",
              "inativa"
            ]
          },
          "permiteAutoCadastro": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "nome",
          "slug",
          "esportes",
          "logoUrl",
          "status",
          "permiteAutoCadastro",
          "createdAt",
          "updatedAt"
        ]
      },
      "EmpresaPaginadaResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmpresaResponseDto"
            }
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 7
          }
        },
        "required": [
          "data",
          "page",
          "pageSize",
          "total"
        ]
      },
      "AdminInicialDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "senha": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          }
        },
        "required": [
          "nome",
          "email",
          "senha"
        ]
      },
      "CreateCompanyDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string"
          },
          "esportes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "adminInicial": {
            "$ref": "#/components/schemas/AdminInicialDto"
          }
        },
        "required": [
          "nome",
          "esportes",
          "adminInicial"
        ]
      },
      "AdminInicialResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string",
            "example": "Ana Souza"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "company_admin"
            ]
          },
          "companyId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "id",
          "nome",
          "email",
          "role",
          "companyId"
        ]
      },
      "EmpresaCriadaResponseDto": {
        "type": "object",
        "properties": {
          "empresa": {
            "$ref": "#/components/schemas/EmpresaResponseDto"
          },
          "adminUsuario": {
            "$ref": "#/components/schemas/AdminInicialResponseDto"
          }
        },
        "required": [
          "empresa",
          "adminUsuario"
        ]
      },
      "AdminDaEmpresaResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string",
            "example": "Ana Souza"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "status": {
            "type": "string",
            "enum": [
              "ativo",
              "inativo"
            ]
          },
          "senhaTemporaria": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "nome",
          "email",
          "status",
          "senhaTemporaria"
        ]
      },
      "ContaDeAdminResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string",
            "example": "Ana Souza"
          },
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "id",
          "nome",
          "email"
        ]
      },
      "SenhaDeAdminResponseDto": {
        "type": "object",
        "properties": {
          "usuario": {
            "$ref": "#/components/schemas/ContaDeAdminResponseDto"
          },
          "senhaTemporaria": {
            "type": "string",
            "example": "pck-Xk4p-9Qm2"
          },
          "expiraEm": {
            "type": "string",
            "format": "date-time"
          },
          "empresaInativa": {
            "type": "boolean"
          }
        },
        "required": [
          "usuario",
          "senhaTemporaria",
          "expiraEm",
          "empresaInativa"
        ]
      },
      "UpdateCompanyDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string"
          },
          "esportes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UpdateCompanyStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ativa",
              "inativa"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "EmpresaPublicaResponseDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Smart Tennis"
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "nome",
          "logoUrl"
        ]
      },
...
      "ChamadaNaoHouveResponseDto": {
        "type": "object",
        "properties": {
          "ocupacaoId": {
            "type": "string",
            "format": "uuid"
          },
          "completude": {
            "type": "string",
            "example": "nao_houve"
          }
        },
        "required": [
          "ocupacaoId",
          "completude"
        ]
      },
...
      "OcorrenciaDaTurmaResponseDto": {
        "type": "object",
        "properties": {
          "ocupacaoId": {
            "type": "string",
            "format": "uuid"
          },
          "data": {
            "type": "string",
            "example": "2026-09-01"
          },
          "horaInicio": {
            "type": "string",
            "example": "18:00"
          },
          "horaFim": {
            "type": "string",
            "example": "19:00"
          },
          "cancelada": {
            "type": "boolean"
          },
          "chamadaFeita": {
            "type": "boolean"
          },
          "marcados": {
            "type": "number",
            "example": 4
          },
          "totalAlunos": {
            "type": "number",
            "example": 8
          },
          "podeLancar": {
            "type": "boolean"
          },
          "estado": {
            "type": "string",
            "enum": [
              "futura",
              "em_andamento",
              "pendente",
              "feita",
              "legada",
              "nao_houve",
              "cancelada"
            ],
            "description": "`futura` = ainda não começou. `em_andamento` = começou e não terminou. `pendente` = terminou sem chamada. `feita` = chamada declarada completa. `legada` = chamada anterior à SPEC-015. `nao_houve` = alguém declarou que a aula não aconteceu (SPEC-030). `cancelada` = ocorrência cancelada."
          }
        },
        "required": [
          "ocupacaoId",
          "data",
          "horaInicio",
          "horaFim",
          "cancelada",
          "chamadaFeita",
          "marcados",
          "totalAlunos",
          "podeLancar",
          "estado"
        ]
      },
      "OcorrenciasDaTurmaPaginadasResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OcorrenciaDaTurmaResponseDto"
            }
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 38
          }
        },
        "required": [
          "data",
          "page",
          "pageSize",
          "total"
        ]
      },
      "LinhaDaChamadaResponseDto": {
        "type": "object",
        "properties": {
          "alunoId": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string",
            "example": "Ana Souza"
          },
          "status": {
            "type": "string",
            "enum": [
              "presente",
              "ausente",
              "justificado"
            ],
            "nullable": true
          },
          "naTurmaHoje": {
            "type": "boolean"
          }
        },
        "required": [
          "alunoId",
          "nome",
          "status",
          "naTurmaHoje"
        ]
      },
      "ChamadaResponseDto": {
        "type": "object",
        "properties": {
          "ocupacaoId": {
            "type": "string",
            "format": "uuid"
          },
          "turmaId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "data": {
            "type": "string",
            "example": "2026-09-01"
          },
          "horaInicio": {
            "type": "string",
            "example": "18:00"
          },
          "horaFim": {
            "type": "string",
            "example": "19:00"
          },
          "cancelada": {
            "type": "boolean"
          },
          "completude": {
            "type": "string",
            "enum": [
              "completa",
              "desconhecida",
              "nao_houve"
            ],
            "nullable": true
          },
          "versao": {
            "type": "string"
          },
          "alunos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LinhaDaChamadaResponseDto"
            }
          }
        },
        "required": [
          "ocupacaoId",
          "turmaId",
          "data",
          "horaInicio",
          "horaFim",
          "cancelada",
          "completude",
          "versao",
          "alunos"
        ]
      },
      "ItemChamadaDto": {
        "type": "object",
        "properties": {
          "alunoId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "presente",
              "ausente",
              "justificado"
            ]
          }
        },
        "required": [
          "alunoId",
          "status"
        ]
      },
      "SalvarChamadaDto": {
        "type": "object",
        "properties": {
          "versao": {
            "type": "string",
            "description": "Opaca: devolva exatamente o que o GET entregou, sem interpretar. O formato já mudou duas vezes (ganhou o cabecalho e a impressao digital da matricula) e pode mudar de novo — quem fizer parse quebra sem aviso. Serve so para comparacao de igualdade.",
            "example": "<opaco — devolva o valor recebido no GET>"
          },
          "itens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemChamadaDto"
            }
          }
        },
        "required": [
          "versao",
          "itens"
        ]
      },
      "ChamadaSalvaResponseDto": {
        "type": "object",
        "properties": {
          "ocupacaoId": {
            "type": "string",
            "format": "uuid"
          },
          "versao": {
            "type": "string"
          },
          "total": {
            "type": "number",
            "example": 8
          }
        },
        "required": [
          "ocupacaoId",
          "versao",
          "total"
        ]
      },
...
    }
  }
}